bindings_test: refactor common api test code
Move common code to api_test.go
